Give every verifier a package and a unique class name
Ten verifiers sat in the default package, and three names were used twice:
four classes called JavaVerifier, and a StaxFeaturesVerifier and an
Xpp3FeaturesVerifier that collided inside the package they already shared.

That worked only because each verifierId directory is compiled on its own
and loaded through its own classloader. It also meant a caller asking for
"JavaVerifier" gave no clue which of the four it wanted, and that the
directories could never be looked at together by anything.

The java and converters verifiers move into
org.codehaus.modello.generator.{java,converters}, following what the other
plugins already do, and the duplicated names gain the prefix that says
which model they verify. The lookups in the tests spell out the full name.

This does not make the tree compilable at build time - the verifiers assert
against classes that do not exist until the generator has run - but the
names no longer depend on the isolation to stay apart.

Generated-by: Claude Opus 5 (1M context)
