+++ modello-plugins/modello-plugin-converters/src/test/verifiers/converters/ConvertersVerifier.java
@@ -1,3 +1,5 @@
+package org.codehaus.modello.generator.converters;
+
 /*
  * Copyright (c) 2004, Codehaus.org
  *
