+++ modello-plugins/modello-plugin-java/src/test/java/org/codehaus/modello/plugin/java/AnnotationsJavaGeneratorTest.java
@@ -54,6 +54,6 @@ public class AnnotationsJavaGeneratorTest extends AbstractModelloJavaGeneratorTe
 
         compileGeneratedSources(8);
 
-        verifyCompiledGeneratedSources("AnnotationsVerifier");
+        verifyCompiledGeneratedSources("org.codehaus.modello.generator.java.AnnotationsVerifier");
     }
 }
