+++ modello-plugins/modello-plugin-java/src/test/java/org/codehaus/modello/plugin/java/BiDirectionalOverrideJavaGeneratorTest.java
@@ -50,6 +50,6 @@ public class BiDirectionalOverrideJavaGeneratorTest extends AbstractModelloJavaG
 
         compileGeneratedSources();
 
-        verifyCompiledGeneratedSources("JavaVerifier");
+        verifyCompiledGeneratedSources("org.codehaus.modello.generator.java.BidirectionalJavaVerifier");
     }
 }
