+++ modello-plugins/modello-plugin-java/src/test/java/org/codehaus/modello/plugin/java/FeaturesJava5GeneratorTest.java
@@ -54,6 +54,6 @@ public class FeaturesJava5GeneratorTest extends AbstractModelloJavaGeneratorTest
 
         compileGeneratedSources("features", 8);
 
-        verifyCompiledGeneratedSources("JavaVerifier");
+        verifyCompiledGeneratedSources("org.codehaus.modello.generator.java.FeaturesJavaVerifier");
     }
 }
