+++ modello-plugins/modello-plugin-java/src/test/java/org/codehaus/modello/plugin/java/FeaturesVersionJavaGeneratorTest.java
@@ -64,6 +64,6 @@ public class FeaturesVersionJavaGeneratorTest extends AbstractModelloJavaGenerat
 
         compileGeneratedSources();
 
-        verifyCompiledGeneratedSources("JavaVerifier");
+        verifyCompiledGeneratedSources("org.codehaus.modello.generator.java.FeaturesVersionJavaVerifier");
     }
 }
