+++ modello-plugins/modello-plugin-java/src/test/java/org/codehaus/modello/plugin/java/JavaGeneratorTest.java
@@ -54,6 +54,6 @@ public class JavaGeneratorTest extends AbstractModelloJavaGeneratorTest {
 
         compileGeneratedSources();
 
-        verifyCompiledGeneratedSources("JavaVerifier");
+        verifyCompiledGeneratedSources("org.codehaus.modello.generator.java.JavaVerifier");
     }
 }
