+++ modello-plugins/modello-plugin-java/src/test/verifiers/annotations/AnnotationsVerifier.java
@@ -1,3 +1,5 @@
+package org.codehaus.modello.generator.java;
+
 import java.lang.annotation.Annotation;
 
 /*
