+++ modello-plugins/modello-plugin-stax/src/test/java/org/codehaus/modello/generator/xml/stax/FeaturesStaxDomGeneratorTest.java
@@ -60,6 +60,6 @@ public class FeaturesStaxDomGeneratorTest extends AbstractModelloJavaGeneratorTe
         // TODO: see why without this, version system property is set to "2.4.1" value after verify
         System.setProperty("version", getModelloVersion());
 
-        verifyCompiledGeneratedSources("org.codehaus.modello.generator.xml.stax.StaxFeaturesVerifier");
+        verifyCompiledGeneratedSources("org.codehaus.modello.generator.xml.stax.StaxFeaturesDomVerifier");
     }
 }
